move buffer_to_texture to Texture::from_buffer
